Revert "Core: Export every XDP action from pythonbpf.helper"
This reverts commit 55fb6d6b9a42947a01917f762a28528ba12a7473.

+++ tests/kernel_selftest_equivalent/xdp/xdp_tx.py
@@ -3,15 +3,18 @@
 # Bounce every packet back out of the interface it arrived on. Upstream is
 # the transmit side of the veth XDP tests.
 #
+# XDP_TX comes from vmlinux because pythonbpf.helper exports only XDP_PASS
+# and XDP_DROP; that is why this lives under vmlinux/.
+
 from pythonbpf import bpf, section, bpfglobal, compile
-from pythonbpf.helper import XDP_TX
+from vmlinux import XDP_TX
 from ctypes import c_void_p, c_int64
 
 
 @bpf
 @section("xdp")
 def xdp_tx(xdp: c_void_p) -> c_int64:
-    return XDP_TX  # bare, like the C: resolved by the return fast path
+    return c_int64(XDP_TX)
 
 
 @bpf
